+++ plonky3/src/lib.rs
@@ -1,5 +1,8 @@
+mod baby_bear;
 mod circuit_builder;
+mod goldilocks;
 mod params;
 mod stark;
+mod stark_babybear;
 
 pub use stark::Plonky3Prover;
